Move ConditionalFormat method from LocalizedString to StringUtilities.
