+++ Eutherion/Shared/Utils/StringUtilities.cs
@@ -19,6 +19,8 @@
 **********************************************************************************/
 #endregion
 
+using System;
+
 namespace Eutherion.Utils
 {
     /// <summary>
@@ -26,6 +28,28 @@ namespace Eutherion.Utils
     /// </summary>
     public static class StringUtilities
     {
+        /// <summary>
+        /// Conditionally formats a string, based on whether or not it has parameters.
+        /// </summary>
+        public static string ConditionalFormat(string localizedString, string[] parameters)
+        {
+            if (parameters == null || parameters.Length == 0) return localizedString;
+
+            try
+            {
+                return string.Format(localizedString, parameters);
+            }
+            catch (FormatException)
+            {
+                // The provided localized format string is in an incorrect format, and/or it contains
+                // more parameter substitution locations than there are parameters provided.
+                // Examples:
+                // string.Format("Test with parameters {invalid parameter}", parameters)
+                // string.Format("Test with parameters {0}, {1} and {2}", new string[] { "0", "1" })
+                return $"{localizedString} {StringUtilities.ToDefaultParameterListDisplayString(parameters)}";
+            }
+        }
+
         /// <summary>
         /// Generates a display string from an array of parameters in the format "({0}, {1}, ...)".
         /// </summary>
