+++ Sandra.UI/AppTemplate/FileLocalizer.cs
@@ -69,7 +69,7 @@ namespace Eutherion.Win.AppTemplate
 
         public override string Localize(LocalizedStringKey localizedStringKey, string[] parameters)
             => Dictionary.TryGetValue(localizedStringKey, out string displayText)
-            ? LocalizedString.ConditionalFormat(displayText, parameters)
+            ? StringUtilities.ConditionalFormat(displayText, parameters)
             : Default.Localize(localizedStringKey, parameters);
 
         private DefaultUIActionBinding switchToLangUIActionBinding;
